+++ infrastructure/url_checker.go
@@ -9,6 +9,7 @@ import (
 	"context"
 	"crypto/tls"
 	"fmt"
+	"github.com/darren/gpac"
 	"log"
 	"net"
 	"net/http"
@@ -41,6 +42,7 @@ type URLCheckResult struct {
 	FetchedAtEpochSeconds int64
 	BodyPatternsFound     []string
 	RemoteAddr            string
+	CheckerTrace          []URLCheckerPluginTrace
 }
 
 // BodyPatternConfig is unmarshalled from the configuration file
@@ -66,6 +68,7 @@ type urlCheckerSettings struct {
 	BodyPatterns          []bodyPattern
 	EnableRequestTracing  bool
 	UrlCheckerPlugins     []string
+	PacScriptURL          string
 }
 
 // URLChecker interface that all layers should conform to
@@ -80,6 +83,7 @@ type URLCheckerClient struct {
 	settings           urlCheckerSettings
 	dnsCache           *cache.Cache
 	checkerPlugins     []URLCheckerPlugin
+	autoProxy          *gpac.Parser
 }
 
 // NewURLCheckerClient instantiates a new basic URL checking client
@@ -91,6 +95,10 @@ func NewURLCheckerClient() *URLCheckerClient {
 		dnsCache: cache.New(defaultCacheExpirationInterval, defaultCacheCleanupInterval),
 	}
 
+	if c.settings.PacScriptURL != "" {
+		c.autoProxy = parsePacScript(c.settings.PacScriptURL)
+	}
+
 	var checkers []URLCheckerPlugin
 
 	// for now, a valid checker may be configured twice, for whatever reason
@@ -98,8 +106,15 @@ func NewURLCheckerClient() *URLCheckerClient {
 		switch checkerName {
 		case "urlcheck":
 			// default client
-			checkers = addChecker(checkers, newLocalURLChecker(c, buildClient(urlCheckerSettings)))
+			checkers = addChecker(checkers, newLocalURLChecker(c, "urlcheck", buildClient(urlCheckerSettings)))
 			log.Println("Added the defaut URL checker")
+			break
+		case "urlcheck-pac":
+			if c.settings.PacScriptURL == "" {
+				panic("Cannot instantiate a 'urlcheck-pac' checkwer without a proxy auto-config script configured")
+			}
+			checkers = addChecker(checkers, newLocalURLChecker(c, "urlcheck-pac", nil))
+			log.Println("Added the PAC file based URL checker")
 			break
 		case "urlcheck-noproxy":
 			// if proxy is defined, add one without the proxy as fallback
@@ -109,7 +124,7 @@ func NewURLCheckerClient() *URLCheckerClient {
 
 			urlCheckerSettingsNoProxy := urlCheckerSettings
 			urlCheckerSettingsNoProxy.ProxyURL = ""
-			checkers = addChecker(checkers, newLocalURLChecker(c, buildClient(urlCheckerSettingsNoProxy)))
+			checkers = addChecker(checkers, newLocalURLChecker(c, "urlcheck-noproxy", buildClient(urlCheckerSettingsNoProxy)))
 			log.Println("Added the URL checker that doesn't use a proxy")
 			break
 		case "_ok_after_1s_on_delay.com":
@@ -121,7 +136,7 @@ func NewURLCheckerClient() *URLCheckerClient {
 				FetchedAtEpochSeconds: 0,
 				BodyPatternsFound:     nil,
 				RemoteAddr:            "",
-			}})
+			}, "_ok_after_1s_on_delay.com"})
 			log.Println("Added the _always_ok checker")
 			break
 		case "_always_ok":
@@ -133,7 +148,7 @@ func NewURLCheckerClient() *URLCheckerClient {
 				FetchedAtEpochSeconds: 0,
 				BodyPatternsFound:     nil,
 				RemoteAddr:            "",
-			}})
+			}, "_always_ok"})
 			log.Println("Added the _always_ok checker")
 			break
 		case "_always_bad":
@@ -145,7 +160,7 @@ func NewURLCheckerClient() *URLCheckerClient {
 				FetchedAtEpochSeconds: 0,
 				BodyPatternsFound:     nil,
 				RemoteAddr:            "",
-			}})
+			}, "_always_bad"})
 			log.Println("Added the _always_bad checker")
 			break
 		default:
@@ -160,6 +175,20 @@ func NewURLCheckerClient() *URLCheckerClient {
 	c.checkerPlugins = checkers
 
 	return c
+}
+
+func parsePacScript(scriptURL string) *gpac.Parser {
+	res, err := resty.New().R().Get(scriptURL)
+	if err != nil {
+		panic(fmt.Errorf("Could not fetch a PAC script from %v: %v", scriptURL, err.Error()))
+	}
+	log.Printf("Read PAC script from %v", scriptURL)
+	script := string(res.Body())
+	pac, err := gpac.New(script)
+	if err != nil {
+		panic(fmt.Errorf("Could not parse the PAC script: %v", err.Error()))
+	}
+	return pac
 }
 
 func addChecker(checkers []URLCheckerPlugin, plugin URLCheckerPlugin) []URLCheckerPlugin {
@@ -187,6 +216,10 @@ func getURLCheckerSettings() urlCheckerSettings {
 			log.Printf("URLCheckerClient is using a proxy: %v", proxyURL)
 			s.ProxyURL = proxyURL
 		}
+	}
+
+	if pacScriptURL := viper.GetString("pacScriptURL"); pacScriptURL != "" {
+		s.PacScriptURL = pacScriptURL
 	}
 
 	s.MaxRedirectsCount = viper.GetUint("HTTPClient.maxRedirectsCount")
@@ -245,6 +278,11 @@ func getURLCheckerSettings() urlCheckerSettings {
 type fakeURLChecker struct {
 	delay        time.Duration
 	alwaysReturn *URLCheckResult
+	name         string
+}
+
+func (l *fakeURLChecker) Name() string {
+	return l.name
 }
 
 func (l *fakeURLChecker) CheckURL(_ctx context.Context, url string, _lastResult *URLCheckResult) (*URLCheckResult, bool) {
@@ -254,42 +292,84 @@ func (l *fakeURLChecker) CheckURL(_ctx context.Context, url string, _lastResult 
 	return l.alwaysReturn, true /* aborts the chain for now */
 }
 
-func newLocalURLChecker(c *URLCheckerClient, client *resty.Client) *localURLChecker {
+func newLocalURLChecker(c *URLCheckerClient, name string, client *resty.Client) *localURLChecker {
 	return &localURLChecker{
 		c:      c,
 		client: client,
+		name:   name,
 	}
 }
 
 type localURLChecker struct {
 	c      *URLCheckerClient
 	client *resty.Client
+	name   string
+}
+
+func (l *localURLChecker) Name() string {
+	return l.name
 }
 
 func (l *localURLChecker) CheckURL(ctx context.Context, urlToCheck string, lastResult *URLCheckResult) (*URLCheckResult, bool) {
-	if lastResult == nil || (lastResult.Code == http.StatusBadGateway) {
-		return l.c.checkURL(ctx, urlToCheck, l.client)
+	if lastResult == nil || shouldRetryBasedOnStatus(lastResult.Code) {
+		client := l.client
+		if client == nil && l.c.settings.PacScriptURL != "" {
+			client = l.autoSelectClientFor(urlToCheck)
+		}
+		if client == nil {
+			panic("cannot instantiate a HTTP client. Please check the configuration")
+		}
+		return l.c.checkURL(ctx, urlToCheck, client)
 	}
 	return lastResult, false
+}
+
+func (l *localURLChecker) autoSelectClientFor(urlToCheck string) *resty.Client {
+	tmpSettings := l.c.settings
+	proxies, err := l.c.autoProxy.FindProxy(urlToCheck)
+	if err == nil && len(proxies) > 0 {
+		// choosing the first available proxy
+		for _, proxy := range proxies {
+			if proxy.Type == "PROXY" {
+				tmpSettings.ProxyURL = proxies[0].URL()
+				break
+			}
+		}
+	} else {
+		log.Printf("Could not find a proxy for %v", urlToCheck)
+	}
+	return buildClient(tmpSettings)
+}
+
+type URLCheckerPluginTrace struct {
+	Name string
+	Code int
 }
 
 // CheckURL checks a single URL
 func (c *URLCheckerClient) CheckURL(ctx context.Context, url string) *URLCheckResult {
 	var lastRes *URLCheckResult = nil
+	var checkerTrace []URLCheckerPluginTrace
 
 	for pos, currentChecker := range c.checkerPlugins {
 		res, shouldAbort := currentChecker.CheckURL(ctx, url, lastRes)
+		checkerTrace = append(checkerTrace, URLCheckerPluginTrace{
+			Name: currentChecker.Name(),
+			Code: res.Code,
+		})
 
 		if pos == 0 && res == nil {
 			panic("first checker should never return nil")
 		}
 
-		if shouldAbort {
-			return res
-		}
-
 		lastRes = res
+
+		if shouldAbort || !shouldRetryBasedOnStatus(lastRes.Code) {
+			break
+		}
 	}
+
+	lastRes.CheckerTrace = checkerTrace
 
 	return lastRes
 }
@@ -361,11 +441,7 @@ func (c *URLCheckerClient) tryGetRequestAndProcessResponseBody(ctx context.Conte
 	var body string
 	// some sites don't allow HEAD requests, try a GET
 	if c.settings.SearchForBodyPatterns ||
-		res.Code == http.StatusForbidden ||
-		res.Code == http.StatusMethodNotAllowed ||
-		res.Code == http.StatusServiceUnavailable ||
-		res.Code == http.StatusNotFound ||
-		res.Code == CustomHTTPErrorCode {
+		shouldRetryBasedOnStatus(res.Code) {
 		response, err := client.R().
 			SetHeader("Accept", c.settings.AcceptHeader).
 			SetContext(ctx).
@@ -381,6 +457,21 @@ func (c *URLCheckerClient) tryGetRequestAndProcessResponseBody(ctx context.Conte
 		res = c.searchForBodyPatterns(res, body)
 	}
 	return res
+}
+
+func shouldRetryBasedOnStatus(code int) bool {
+	if code < 300 {
+		return false
+	}
+
+	return code == http.StatusForbidden ||
+		code == http.StatusMethodNotAllowed ||
+		code == http.StatusServiceUnavailable ||
+		code == http.StatusNotFound ||
+		code == CustomHTTPErrorCode ||
+		code == http.StatusGatewayTimeout ||
+		code == http.StatusBadGateway ||
+		code == http.StatusRequestTimeout
 }
 
 func (c *URLCheckerClient) tryHeadRequestDefault(ctx context.Context, urlToCheck string, client *resty.Client) *URLCheckResult {