+++ infrastructure/url_checker_plugin.go
@@ -4,6 +4,9 @@ import "context"
 
 // URLCheckerPlugin represents one low-level URL checker in a chain of checkers
 type URLCheckerPlugin interface {
+	// Name returns the name of the plugin to use in logging and result reporting
+	Name() string
+
 	// CheckURL gets the urlToCheck and lastResult, which it can process, and return the next result
 	// and a boolean flag, whether the chain should be interrupted, and the last result - simply returned
 	// ctx can be used to cancel the request prematurely